Add test pings
