+++ pages/src/App.js
@@ -30,12 +30,15 @@ class App extends Component {
       console.log("Message received from server");
       console.log(event.data);
     });
-    this.heartBeatIntervalId = setInterval(() => this.websocket?.send('{"type": "PING"}'), 3000)
+    this.heartBeatIntervalId = setInterval(
+      () => this.websocket?.send('{"type": "PING"}'),
+      3000
+    );
   }
 
   componentWillUnmount() {
     this.websocket?.close();
-    clearInterval(this.heartBeatIntervalId)
+    clearInterval(this.heartBeatIntervalId);
   }
 
   render() {
@@ -48,10 +51,10 @@ class App extends Component {
         </button>
         <button
           onClick={() => {
-            this.websocket?.send("ping");
+            this.websocket?.send(JSON.stringify({ type: "TEST" }));
           }}
         >
-          Ping
+          Test
         </button>
       </div>
     );
