+++ workers/src/index.js
@@ -1,24 +1,3 @@
-function handleWebSocket(request) {
-  const upgradeHeader = request.headers.get("Upgrade");
-  if (!upgradeHeader || upgradeHeader !== "websocket") {
-    return new Response("Expected Upgrade: websocket", { status: 426 });
-  }
-
-  const webSocketPair = new WebSocketPair();
-  const [client, server] = Object.values(webSocketPair);
-
-  server.accept();
-  new HeartBeatManager({}, server);
-  server.addEventListener("message", (event) => {
-    console.log(event.data);
-    server.send("PONG");
-  });
-
-  return new Response(null, {
-    status: 101,
-    webSocket: client,
-  });
-}
 export { GameRoom } from "./GameRoom";
 
 // eslint-disable-next-line import/no-anonymous-default-export
@@ -35,53 +14,6 @@ export default {
     }
   },
 };
-
-export class HeartBeatManager {
-  constructor(
-    game,
-    webSocket,
-    checkMilliseconds = 3000,
-    timeoutMilliseconds = 30000
-  ) {
-    this.game = game;
-    this.webSocket = webSocket;
-    this.checkMilliseconds = checkMilliseconds;
-    this.timeoutMilliseconds = timeoutMilliseconds;
-    this.heartBeatLastReceivedTime = new Date().getTime();
-
-    this.checkIntervalID = setInterval(
-      () => this.heartBeatCheck(),
-      checkMilliseconds
-    );
-
-    this.webSocket.addEventListener("message", (event) => {
-      try {
-        let message = JSON.parse(event.data);
-        if (message["type"] === "PING") {
-          this.heartBeatReceived();
-        }
-      } catch (e) {}
-    });
-
-    this.webSocket.addEventListener("close", (event) =>
-      clearInterval(this.checkIntervalID)
-    );
-  }
-
-  heartBeatReceived() {
-    console.log(`Heartbeat received from ${this.webSocket}`);
-    this.heartBeatLastReceivedTime = new Date().getTime();
-    // this.game.markPlayerAlive();
-  }
-
-  heartBeatCheck() {
-    let now = new Date().getTime();
-    if (now - this.heartBeatLastReceivedTime > this.timeoutMilliseconds) {
-      // this.game.markPlayerDead();
-      console.log(`${this.webSocket} is dead`);
-    }
-  }
-}
 
 async function handleApiRequest(path, request, env) {
   switch (path[0]) {
